requirements.txt
gitpython
pulumi
pulumi-docker
pulumi-gcp
pulumi-kubernetes
pyyaml
